Update guest view in layout file
The existing block of text in the guest view section of the layout file has been replaced with a succinct request to register for unlocking options. It also includes 'login' and 'register' buttons linked to their respective pages. This change makes the user interface more direct and actionable for guests.