Add Global Team seeding and role assignment enhancements
This commit introduces an additional team seeding functionality in the 'DatabaseSeeder.php' and 'RoleAndPermissionSeeder.php' files. A new TeamSeeder class is created to seed a 'Global Team'. Furthermore, an extra setPermissionsTeamId line is added on RoleAndPermissionSeeder to set permissions for the team. Additionally, a user is now assigned both the admin and support roles. The term 'Global Team' is also localized in the 'fa.json' file.