Add Livewire configuration and user index views
Configured the Livewire settings like the namespace, view path, layout, and other settings to facilitate the creation of Livewire components in the application. Also added a new route for the user index page that uses a Livewire component. Updated the user panel in the app layout and added views for the admin user index page.