Update Laravel and Symfony packages versions
This commit updates the versions of several Laravel and Symfony related packages. Each updated package includes not only the newer version but also the updated reference and time, ensuring the most current and secure versions of these packages are used. This helps to improve performance, security, and functionality of the overall application.