Update default font to Vazirmatn
This commit changes the default font from Tahoma to Vazirmatn. The Vazirmatn font has been added as a dependency in the package.json and appropriately imported in the CSS files. All elements now use this as the default font as per the updates in the base layer of the CSS.